requirements.txt
psycopg2-binary==2.7.7
pandas==0.24.2
pygsheets==2.0.0